prototype of sending led data function

+++ include/led_memory.h
@@ -2,6 +2,6 @@
 #define LED_MEMORY
 
 #include <cstdint>
-extern std::uint8_t led_memory[8][8];
+extern std::uint32_t led_memory[8][8];
 
 #endif // LED_MEMORY
