+++ src/led_memory.cpp
@@ -1,3 +1,3 @@
 #include <cstdint>
 
-std::uint8_t led_memory[8][8] {0};
+std::uint32_t led_memory[8][8] {0};
